requirements.txt
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
faiss-cpu==1.10.0
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
fonttools==4.57.0
fsspec==2025.3.2
h11==0.14.0
idna==3.10
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-engineio==4.12.0
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==79.0.0
simple-websocket==1.1.0
simsimd==6.2.1
six==1.17.0
stringzilla==3.12.5
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.114
ultralytics-thop==2.0.14
urllib3==2.4.0
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
